refactor: :recycle: add namespace Modules
- split all the modules from namespace Interface to GameClass
- split ModuleFactory
